requirements.txt
openai
Pillow
fastapi
uvicorn
opencv-python
matplotlib